### 2023-04-14 #5367 @arfon
Note this is a resubmission of https://github.com/openjournals/joss-reviews/issues/5032
