### 2023-04-14 #5367 @arfon
@SimonMolinsky – please quote the affiliations on [this line](https://github.com/nokaut/wsknn/blob/paper/paper/paper.md?plain=1#L12) (i.e., `"1, 2"`). That should fix your paper compilation.
