### 2023-05-02 #5367 @AoifeHughes
Hi @SimonMolinsky, happy to be involved. First things first, let's find some reviewers. Do you have anyone (or two) in mind who would be appropriate to review? You can use: https://reviewers.joss.theoj.org/reviewers. Or you can suggest others and I can reach out to them. Otherwise, I can start contacting some potential reviewers.
