### 2023-05-03 #5367 @SimonMolinsky
Hi @AoifeHughes ,

I've done some research, and here's the subjectively ordered list:

- https://reviewers.joss.theoj.org/reviewers/1866 (recommender systems),
- https://reviewers.joss.theoj.org/reviewers/2250 (behavioral and social sciences),
- https://reviewers.joss.theoj.org/reviewers/2602 (Computational Social Science),
- https://reviewers.joss.theoj.org/reviewers/1198 (General knowledge about ML / AI).

Anyone with general knowledge about ML models could check the package. It could help if the reviewer has experience related to the following:

- recommender systems,
- sequences of events,
- e-commerce & retail.
