### 2023-05-24 #5367 @inpefess
Hi @AoifeHughes, no worries, I prefer GitHub notifications to emails for the JOSS. Yes, I'd be happy to review.
