+++ wsknn/preprocessing/static_parsers/json_parser/parse.py
@@ -1,0 +1,68 @@
+import gzip
+from typing import Dict
+
+
+def parse_gizipped(dataset: str,
+                   allowed_actions: Dict,
+                   purchase_action_name: str,
+                   session_id_key: str,
+                   product_key: str,
+                   action_key: str,
+                   time_key: str,
+                   time_to_numeric: bool,
+                   time_to_datetime: bool,
+                   datetime_format: str):
+    """
+    Function parses given gzipped JSONL file into Sessions and Items objects.
+
+    Parameters
+    ----------
+    dataset : str
+        The gzipped JSONL file with events.
+
+    allowed_actions : Dict, optional
+        Allowed actions and their weights.
+
+    purchase_action_name: Any, optional
+        The name of the final action (it is required to apply weight into the session vector).
+
+    session_id_key : str
+        The name of the session key.
+
+    product_key : str
+        The name of the product key.
+
+    action_key : str
+        The name of the event action type key.
+
+    time_key : str
+        The name of the event timestamp key.
+
+    time_to_numeric : bool, default = True
+        Transforms input timestamps to float values.
+
+    time_to_datetime : bool, default = False
+        Transforms input timestamps to datatime objects. Setting ``datetime_format`` parameter is required.
+
+    datetime_format : str
+        The format of datetime object.
+
+    Returns
+    -------
+    parsed_items, parsed_sessions : Items, Sessions
+        The mappings of item-session and session-items.
+    """
+    with gzip.open(dataset, 'rt', encoding='UTF-8') as unzipped_f:
+        parsed_items, parsed_sessions = parse_fn(
+            unzipped_f,
+            allowed_actions,
+            purchase_action_name,
+            session_id_key,
+            product_key,
+            action_key,
+            time_key,
+            time_to_numeric,
+            time_to_datetime,
+            datetime_format)
+
+    return parsed_items, parsed_sessions
