+++ wsknn/preprocessing/parse_static.py
@@ -8,6 +8,54 @@ from wsknn.preprocessing.structure.item import Items
 from wsknn.preprocessing.structure.session import Sessions
 
 
+def _clean_times(times: Any,
+                 time_to_numeric: bool,
+                 time_to_datetime: bool,
+                 datetime_format: str) -> List:
+    """
+    Function cleans timestamps if they don't have valid data type.
+
+    Parameters
+    ----------
+    times : List
+        The list with timestamps.
+
+    time_to_numeric : bool
+        Should time be transformed to numeric?
+
+    time_to_numeric : bool, default = True
+        Transforms input timestamps to float values.
+
+    time_to_datetime : bool, default = False
+        Transforms input timestamps to datatime objects. Setting `datetime_format` parameter is required.
+
+    datetime_format : str
+        The format of datetime object.
+
+    Returns
+    -------
+    : List
+        Parsed times.
+    """
+
+    parsed_times = times
+
+    if isinstance(times, List):
+        if time_to_numeric:
+            parsed_times = [float(x) for x in times]
+
+        if time_to_datetime:
+            parsed_times = [datetime.strptime(x, datetime_format) for x in times]
+    else:
+        if time_to_numeric:
+            parsed_times = float(times)
+
+        if time_to_datetime:
+            parsed_times = datetime.strptime(times, datetime_format)
+
+    return parsed_times
+
+
 def is_user_item_interaction(e_action: str, allowed_actions: Iterable) -> bool:
     """
     Checks if event has a valid action type.
@@ -112,72 +160,6 @@ def parse_files(dataset: str,
     return items, sessions
 
 
-def parse_gz_fn(dataset: str,
-                allowed_actions: Dict,
-                purchase_action_name: str,
-                session_id_key: str,
-                product_key: str,
-                action_key: str,
-                time_key: str,
-                time_to_numeric: bool,
-                time_to_datetime: bool,
-                datetime_format: str):
-    """
-    Function parses given gzipped JSONL file into Sessions and Items objects.
-
-    Parameters
-    ----------
-    dataset : str
-        The gzipped JSONL file with events.
-
-    allowed_actions : Dict, optional
-        Allowed actions and their weights.
-
-    purchase_action_name: Any, optional
-        The name of the final action (it is required to apply weight into the session vector).
-
-    session_id_key : str
-        The name of the session key.
-
-    product_key : str
-        The name of the product key.
-
-    action_key : str
-        The name of the event action type key.
-
-    time_key : str
-        The name of the event timestamp key.
-
-    time_to_numeric : bool, default = True
-        Transforms input timestamps to float values.
-
-    time_to_datetime : bool, default = False
-        Transforms input timestamps to datatime objects. Setting ``datetime_format`` parameter is required.
-
-    datetime_format : str
-        The format of datetime object.
-
-    Returns
-    -------
-    parsed_items, parsed_sessions : Items, Sessions
-        The mappings of item-session and session-items.
-    """
-    with gzip.open(dataset, 'rt', encoding='UTF-8') as unzipped_f:
-        parsed_items, parsed_sessions = parse_fn(
-            unzipped_f,
-            allowed_actions,
-            purchase_action_name,
-            session_id_key,
-            product_key,
-            action_key,
-            time_key,
-            time_to_numeric,
-            time_to_datetime,
-            datetime_format)
-
-    return parsed_items, parsed_sessions
-
-
 def parse_jsonl_fn(dataset: str, allowed_actions: Dict, purchase_action_name: str, session_id_key: str,
                    product_key: str,
                    action_key: str,
@@ -244,180 +226,3 @@ def parse_jsonl_fn(dataset: str, allowed_actions: Dict, purchase_action_name: st
         )
 
[…]
