### 2023-07-10 #5639 @editorialbot
[REVIEW]: WSKNN - Weighted Session-based K-NN recommender system
**Submitting author:** @SimonMolinsky (<a href="http://orcid.org/0000-0003-3525-2104">Szymon Moliński</a>)
**Repository:** https://github.com/nokaut/wsknn
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.1
