### 2023-07-10 nokaut/wsknn#29 @inpefess
Can't run example from the README
The README example starts with

```python
ITEMS = 'demo-data/items.pkl'
SESSIONS = 'demo-data/sessions.pkl'
```

what are these `pkl` files, how can a potential new user find or generate them? There is a `demo-data` folder, but not in the root, and there are no `items.pkl` or `sessions.pkl` files anywhere in the repository.
