### 2023-07-10 #5639 @inpefess
https://github.com/nokaut/wsknn/issues/29 (blocker for "Example usage")
