### 2023-07-10 nokaut/wsknn#29 @SimonMolinsky
Hi @inpefess

Thanks for pointing this out! I have created demo examples here: https://github.com/nokaut/wsknn/tree/main/demo-notebooks, and I see that `README` content differs from those. I will update `README` accordingly to use *recsys* data.
