### 2023-07-10 nokaut/wsknn#30 @inpefess
No link to documentation from GitHub page
There are quality documentation pages hosted on RTD, and they are easily accessible from the PyPI page of the package, but a person who discovered the project through GitHub might get the wrong impression that there are no readily accessible docs. Please add the link to the documentation to README.
