### 2023-07-10 #5639 @inpefess
https://github.com/nokaut/wsknn/issues/30 (Functionality documentation)
