### 2023-07-11 #5639 @inpefess
Please find my suggestions to improve the paper here: https://github.com/nokaut/wsknn/issues/32
Thank you for creating and sharing the WSKNN! Unfortunately, we still don't have enough well-documented and maintained recommender system packages in free access, so your efforts are most laudable! I'm looking forward to your corrections and further discussion.
