+++ README.md
@@ -6,7 +6,11 @@
 
 Do you build a recommender system for your website? K-nearest neighbors algorithm is a good choice if you are looking for a simple, fast, and explainable solution. Weighted-session-based k-nn recommendations are close to the state-of-the-art, and we don't need to tune multiple hyperparameters and build complex deep learning models to achieve a good result.
 
-### How does it work?
+## Documentation
+
+API Documentation is available here: [WSKNN Docs](https://wsknn.readthedocs.io/en/latest/index.html)
+
+## How does it work?
 
 You provide two input structures as **training** data:
 
@@ -44,7 +48,7 @@ The package is lightweight. It depends only on the `numpy` and `pyyaml`.
 Moreover, we can provide a package for non-programmers, and they can use `settings.yaml` to control a model behavior.
 
 
-### Why should we use WSKNN?
+## Why should we use WSKNN?
 
 - training is faster than deep learning or XGBoost algorithms, model memorizes map of session-items and item-sessions,
 - recommendations are easy to control. We can change how the algorithm works in just a few lines... of text,
@@ -54,14 +58,14 @@ Moreover, we can provide a package for non-programmers, and they can use `settin
 
 The model was created along with multiple other approaches: based on RNN (GRU/LSTM), matrix factorization, and others. Its performance was always very close to the level of fine-tuned neural networks, but it was much easier and faster to train.
 
-### What are the limitations of WSKNN?
+## What are the limitations of WSKNN?
 
 - model memorizes session-items and item-sessions maps, and if your product base is large and you use sessions for an extended period, then the model may be too big to fit an available memory; in this case, you can 
 categorize products and train a different model for each category,
 - response time may be slower than from other models, especially if there are available many sessions,
 - there's additional overhead related to the preparation of the input.
 
-### Example
+## Example
 
 Example below is available in `demo-notebooks/demo-readme.ipynb` notebook.
 
@@ -114,10 +118,6 @@ Item: 214561775 | weight: 1.2478940715049494
 Item: 214821020 | weight: 1.2478940715049494
 Item: 214848322 | weight: 1.2478940715049494
 ```
-
-## Preprocessing Stage
-
-
 
 ## Setup
 
