### 2023-07-18 nokaut/wsknn#35 @inpefess
Contribution guide
https://github.com/nokaut/wsknn/blob/main/CONTRIBUTING.md looks good, but I still can propose several improvements:

1. add a link from README to the contribution guide (it's easy to overlook since there are many other files in the root)
2. provide more detailed instructions on setting up the developer's environment (let's say I've got my virtualenv/conda env, how do I install all development requirements? There is a `requirements.txt`, but there are also `setup.cfg` and `pyproject.toml`, it's better to be precise)
3. indicate that one has to run `pytest` only from `tests` directory (when run from the project root, it fails because of not finding `tdata` folder)
4. the guide reads: "Make sure your code lints." How should one do that? (from the GitHub Actions it's obvious it means `flake8`, but from the guide, it doesn't follow)
5. the guide reads: "Use a PEP8 Guidelines". Can one do it automatically? (the same as the previous point, I guess)
