### 2023-07-18 #5639 @inpefess
https://github.com/nokaut/wsknn/issues/35 (Community guidelines)
