### 2023-07-18 nokaut/wsknn#37 @inpefess
An example of loading tabular data and a bug in the `parse_files` function
Several popular papers on sequential recommenders (e.g. BERT4Rec and SASRec) rely on tabular data (MovieLens, Amazon, Steam). I tried to run WSKNN on MovieLens 25M and failed to apply the `parse_files` function. The argument `allowed_actions` is `None` by default, so I didn't pass it, but then inside the function presumes it's not-None and fails. It wouldn't be unreasonable to assume that all the actions appearing in the `action_key` field are allowed by default and work with the omitted `allowed_actions` dictionary gracefully. And, of course, the usage example with a popular open dataset instead of the package-specific one might make it even more user-friendly.
