### 2023-07-18 #5639 @inpefess
https://github.com/nokaut/wsknn/issues/37 (Functionality: Have the functional claims of the software been confirmed?)

Without an example on tabular data, it seems to be relatively hard to use the package for e-commerce applications since a server log (or a transactions table) of the format

timestamp,session_id,item_id,action_type

is still a wide-spread data representation scheme in this field (to my best knowledge).
