### 2023-07-19 #5639 @SimonMolinsky
@inpefess thank you for your thorough review and comments! Those are invaluable for the package, and the requested changes will make it better for the community. I will start working on it, and let you know when every step is done.
