+++ paper/paper.bib
@@ -74,3 +74,21 @@ url={https://doi.org/10.1007/s11257-018-9209-6}
   journal = {GitHub repository},
   url = {https://github.com/nokaut/wsknn/blob/main/demo-notebooks/demo.ipynb}
 }
+
+@inproceedings{recsys2015data,
+author = {Ben-Shimon, David and Tsikinovsky, Alexander and Friedmann, Michael and Shapira, Bracha and Rokach, Lior and Hoerle, Johannes},
+title = {RecSys Challenge 2015 and the YOOCHOOSE Dataset},
+year = {2015},
+isbn = {9781450336925},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/2792838.2798723},
+doi = {10.1145/2792838.2798723},
+abstract = {The 2015 ACM Recommender Systems Challenge offered the opportunity to work on a large-scale e-commerce dataset from a big retailer in Europe which is accepting recommender system as a service from YOOCHOOSE. Participants tackled the problem of predicting what items a user intends to purchase, if any, given a click sequence performed during an activity session on the e-commerce website. The challenge ran for seven months and was very successful, attracting 850 teams from 49 countries which submitted a total of 5,437 solutions. The winners of the challenge scored approximately 50% of the maximum score, which we considered as an impressive achievement. In this paper we provide a brief overview of the challenge and its results.},
+booktitle = {Proceedings of the 9th ACM Conference on Recommender Systems},
+pages = {357–358},
+numpages = {2},
+keywords = {recsys challenge 2015, e-commerce, yoochoose, recommender systems},
+location = {Vienna, Austria},
+series = {RecSys '15}
+}
