### 2023-08-02 8b810d1
performance metrics updates
