### 2023-08-02 nokaut/wsknn#31 @SimonMolinsky
Ok, I've done this part. There are some updates in the paper, and in the `README`. I've created an additional notebook with computational performance measurements.
