+++ requirements.txt
@@ -1,8 +1,3 @@
 numpy
 pyyaml
 more_itertools
-# [dev]
-# matplotlib
-# setuptools
-pytest
-tqdm
