### 2023-08-10 1e3a378
Create session_to_item_map.py
