### 2023-08-10 361b30f
create item-sessions map from session-items map
