### 2023-08-10 nokaut/wsknn#37 @SimonMolinsky
Thanks, this bug seems to be critical. I will perform some tests on the tabular datasets mentioned by you too, and when everything works fine, I'll close this issue.
