### 2023-08-14 #5639 @inpefess
@AoifeHughes, reviewer instructions command "to complete your review in the next six weeks", so the deadline is today. I've finished the first iteration, and the author did well revising the code and the paper. I will need to read a new version one more time after the author fixes two last issues (https://github.com/nokaut/wsknn/issues/35 and https://github.com/nokaut/wsknn/issues/37). @SimonMolinsky many thanks for improving the package!
