+++ wsknn/__init__.py
@@ -1,9 +1,7 @@
-from wsknn.fit_transform import fit
+from wsknn.fit_transform import fit, Items, Sessions
 from wsknn.predict import predict
 from wsknn.model.wsknn import WSKNN
 from wsknn.preprocessing.parse_static import parse_files
-from wsknn.preprocessing.structure.item import Items
-from wsknn.preprocessing.structure.session import Sessions
 
 
 __version__ = '1.1.1'
