+++ wsknn/preprocessing/static_parsers/pandas_parser.py
@@ -1,0 +1,197 @@
+from typing import Dict, List
+import pandas as pd
+from wsknn.preprocessing.structure.session_to_item_map import map_sessions_to_items
+
+
+def _clip_multiple_transactions(session, items, times, actions, locs, min_session_length):
+    loclen = len(items) - 1
+
+    parsed = []
+    pidx = None
+
+    for idx, lc in enumerate(locs):
+        if lc == 0:
+            continue
+        else:
+            if idx == 0:
+                previous = 0
+            else:
+                previous = locs[idx-1] + 1
+
+            if lc + 1 < loclen:
+                lc = lc + 1
+
+            nitem = items[previous:lc]
+            ntime = times[previous:lc]
+            nact = actions[previous:lc]
+            if len(nitem) >= min_session_length:
+                parsed.append(
+                    [session, nitem, ntime, nact]
+                )
+        pidx = idx
+
+    if pidx < loclen:
+        nitem = items[pidx:]
+        ntime = times[pidx:]
+        nact = actions[pidx:]
+        if len(nitem) >= min_session_length:
+            parsed.append(
+                [session, nitem, ntime, nact]
+            )
+
+    return parsed
+
+
+def _parse_full_ds(sess_idx, items, times, actions, is_transaction, min_session_length):
+    if is_transaction is None:
+        # Skip dividing sessions and put everything in one sequence
+        parsed = {
+            x: [items[idx], times[idx], actions[idx]]
+            for idx, x in enumerate(sess_idx) if len(items[idx]) >= min_session_length}
+    else:
+        parsed = {}
+        for idx, x in enumerate(is_transaction):
+            if 1 in x:
+                if len(items[idx]) > min_session_length:
+                    parsed[sess_idx[idx]] = [items[idx], times[idx], actions[idx]]
+            # TODO: complex parsing schema, where single session may occur multiple times! FUTURE
+            #     # We have a transaction here, time to divide sequence
+            #     t_idx = list(locate(x))
+            #     # Divide sequences based on the transaction indices
+            #     parsed.extend(
+            #         _clip_multiple_transactions(
+            #             session=sess_idx[idx], items=items[idx], times=times[idx], actions=actions[idx], locs=t_idx,
+            #             min_session_length=min_session_length
+            #         )
+            #     )
+            # else:
+            #     if len(items[idx]) >= min_session_length:
+            #         parsed.append(
+            #             {sess_idx[idx]: [items[idx], times[idx], actions[idx]]}
+            #         )
+    return parsed
+
+
+def _prepare_values_session_map(df: pd.DataFrame,
+                                session_id_key: str,
+                                product_key: str,
+                                time_key: str,
+                                action_key: str = None,
+                                purchase_action_name: str = None):
+    df = df.sort_values([session_id_key, time_key])
+    gdf = df.groupby(session_id_key)
+    timestamps = gdf[time_key].apply(list).values
+    products = gdf[product_key].apply(list)
+    actions = None
+    transactions = None
+    if action_key is not None:
+        actions = gdf[action_key].apply(list).values
+    if purchase_action_name is not None:
+        transactions = gdf['is_transaction'].apply(list).values
+
+    return products.index, products.values, timestamps, actions, transactions
+
+
+def _build_maps_from_df(df: pd.DataFrame,
+                        session_id_key: str,
+                        product_key: str,
+                        time_key: str,
+                        action_key: str = None,
+                        purchase_action_name: str = None,
+                        min_session_length: int = 3) -> Dict:
+    # Prepare data for session map
+    sess_idx, items, times, actions, is_transaction = _prepare_values_session_map(
+        df,
+        session_id_key,
[…]
