+++ wsknn/__init__.py
@@ -2,6 +2,7 @@ from wsknn.fit_transform import fit, Items, Sessions
 from wsknn.predict import predict
 from wsknn.model.wsknn import WSKNN
 from wsknn.preprocessing.parse_static import parse_files
+from wsknn.preprocessing.static_parsers.pandas_parser import parse_pandas
 
 
-__version__ = '1.1.1'
+__version__ = '1.2.0'
