+++ docs/requirements.txt
@@ -1,4 +1,5 @@
 numpy
+pandas
 pyyaml
 sphinx
 numpydoc
