### 2023-08-28 nokaut/wsknn#37 @SimonMolinsky
@inpefess It took me longer than I expected because the preprocessing time of the 25M MovieLens dataset was a disaster with core features. I've written a parser based on the `pandas` package, and it processes data in a reasonable time. I've created two additional tutorials based on the MovieLens datasets. I've removed the error. Now, I'm wrapping things up. I will publish the new release with all those features. I must change descriptions slightly so that I will do it along with the paper and contributor's guide corrections. I'll let you know when everything's ready in the JOSS review thread.
