+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 # Contribution to WSKNN
 
-We love your input! We want to make contributing to this project as easy and transparent as possible, whether it's:
+We love your input! We want to make contributing to this project as easy and transparent as possible, whether it's
 
 * Reporting a bug
 * Discussing the current state of the code
@@ -10,9 +10,76 @@ We love your input! We want to make contributing to this project as easy and tra
 
 ## Where should I start?
 
-Here, on **GitHub**! We use **GitHub** to host the code, to track issues and feature requests, as well as accept pull requests.
+Here, on **GitHub**! We use **GitHub** to host the code, track issues, and feature requests, and accept pull requests.
 
 ---
+
+## Developer setup
+
+Setup for developers differs from the package installation from `PyPI`.
+
+1. Fork the `wsknn` repository.
+2. Clone forked repository.
+3. Connect the main repository with your fork locally:
+
+```shell
+git remote add upstream https://github.com/nokaut/wsknn.git
+
+```
+
+4. Synchronize your repository with the core repository.
+
+
+```shell
+git checkout main
+git pull upstream main
+
+```
+
+5. Create your branch.
+
+```shell
+git checkout -b name-of-your-branch
+
+```
+
+6. Create [virtual environment](https://docs.python.org/3/library/venv.htmlc) or [conda environment](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-with-commands).
+7. Activate your environment.
+8. Install requirements listed in the `requirements-dev.txt` file.
+
+**Virtual Environment**
+
+```shell
+>> (your-virtual-environment) pip install -r requirements-dev.txt
+         
+```
+
+**Conda**
+
+```shell
+>> (your-conda-environment) conda install -c conda-forge --file requirements-dev.txt
+
+```
+
+9. Make changes in a code or write something new.
+10. Write tests if required.
+11. Perform tests with `pytest`. (Run tests from the `tests` directory).
+
+```shell
+>> (your-environment) (your-username:~/path/wsknn/tests) pytest
+
+```
+
+12. If all tests pass, push changes into your fork.
+
+```shell
+git add .
+git commit -m "description what you have done"
+git push origin name-of-your-branch
+
+```
+
+13. Navigate to your repository. You should see a button `Compare and open a pull request`. Use it to make a pull request! Send it to the `dev` branch in the main repository. **Don't send pull requests into `main` branch of the core repository!**
 
 ## We Use [Github Flow](https://guides.github.com/introduction/flow/index.html), So All Code Changes Happen Through Pull Requests
 Pull requests are the best way to propose changes to the codebase (we use [Github Flow](https://guides.github.com/introduction/flow/index.html)). We actively welcome your pull requests:
@@ -20,9 +87,9 @@ Pull requests are the best way to propose changes to the codebase (we use [Githu
 1. Fork the repo and create your branch from `main`.
 2. If you've added code that should be tested, add tests in the `test` package. We use Python's `pytest` package to perform testing.
 3. If you've changed APIs, update the documentation.
-4. Ensure the test suite passes.
-5. Make sure your code lints.
-6. Issue that pull request!
+4. Ensure the test suite passes. Run tests from `tests` directory, otherwise you will encounter an error.
+5. Make sure your code lints, you can use `flake8` or linters included in a development software (*Pycharm*). Linters check [PEP8 Python Guidelines](https://www.python.org/dev/peps/pep-0008/), and it is recommended to read them first.
+6. Issue that pull request.
 
 ## Any contributions you make will be under the BSD 3-Clause "New" or "Revised" License
[…]
