+++ README.md
@@ -4,7 +4,7 @@
 
 ## Weighted session-based k-NN - Intro
 
-Do you build a recommender system for your website? K-nearest neighbors algorithm is a good choice if you are looking for a simple, fast, and explainable solution. Weighted-session-based k-nn recommendations are close to the state-of-the-art, and we don't need to tune multiple hyperparameters and build complex deep learning models to achieve a good result.
+Do you build a recommender system for your website? K-nearest neighbors algorithm is a good choice if you are looking for a simple, fast, and explainable solution. Weighted-session-based k-nn recommendations are [close to the state-of-the-art](#sknn-performance), and we don't need to tune multiple hyperparameters and build complex deep learning models to achieve a good result.
 
 ## Documentation
 
@@ -56,7 +56,7 @@ Moreover, we can provide a package for non-programmers, and they can use `settin
 - swift prototyping,
 - easy to run in production.
 
-The model was created along with multiple other approaches: based on RNN (GRU/LSTM), matrix factorization, and others. Its performance was always very close to the level of fine-tuned neural networks, but it was much easier and faster to train.
+The model was created along with multiple other approaches: based on RNN (GRU/LSTM), matrix factorization, and others. [Its performance was always very close to the level of fine-tuned neural networks](#comparison-between-dl-and-wsknn), but it was much easier and faster to train.
 
 ## What are the limitations of WSKNN?
 
@@ -131,10 +131,16 @@ It works with Python versions greater or equal to 3.8.
 
 ## Requirements
 
-| Package Version | Python versions | Requirements                  |
-|-----------------|-----------------|-------------------------------|
-| 0.1.x           | 3.6+            | numpy, pyyaml                 |
- | 1.x             | 3.8+            | numpy, pyyaml, more_itertools |
+| Package Version | Python versions | Requirements                                |
+|-----------------|-----------------|---------------------------------------------|
+| 0.1.x           | 3.6+            | numpy, pyyaml                               |
+| 1.1.x           | 3.8+            | numpy, more_itertools, pyyaml               |
+| 1.2.x           | 3.8+            | numpy, more_itertools, pandas, pyyaml, tqdm |
+
+## Contribution
+
+We welcome all submissions, issues, feature requests, and bug reports! To learn how to contribute to the package please visit [CONTRIBUTION.md file](https://github.com/nokaut/wsknn/blob/main/CONTRIBUTING.md)
+
 
 ## Developers
 
@@ -154,53 +160,59 @@ Szymon Moliński. (2022). WSKNN - Weighted Session-based k-NN Recommendations in
 
 - Twardowski, B., Zawistowski, P., Zaborowski, S. (2021). Metric Learning for Session-Based Recommendations. In: Hiemstra, D., Moens, MF., Mothe, J., Perego, R., Potthast, M., Sebastiani, F. (eds) Advances in Information Retrieval. ECIR 2021. Lecture Notes in Computer Science(), vol 12656. Springer, Cham. https://doi.org/10.1007/978-3-030-72113-8_43
 
+### SKNN performance
+
+The article compares performance of mutiple session-based recommender systems.
+
+- Ludewig, M., Jannach, D. Evaluation of session-based recommendation algorithms. User Model User-Adap Inter 28, 331–390 (2018). https://doi.org/10.1007/s11257-018-9209-6
+
 ## Funding
 
-![Funding](./eu_funding_logos/FE_POIR_poziom_engl-1_rgb.jpg)
+![Funding](./imgs/eu_funding_logos/FE_POIR_poziom_engl-1_rgb.jpg)
 
 - Development of the package was partially based on the research project
 **E-commerce Shopping Patterns Prediction System** that 
 was founded under Priority Axis 1.1 of Smart Growth Operational Programme 2014-2020 for Poland
[…]
