+++ setup.cfg
@@ -34,6 +34,8 @@ install_requires =
     numpy
     more_itertools
     pyyaml
+    tqdm
+    pandas
 [options.packages.find]
 exclude =
     tests*
