### 2023-09-01 fc5767c
Paper and bibliography updates
