+++ paper/paper.md
@@ -22,37 +22,90 @@ bibliography: paper.bib
 
 # Summary
 
-The users of e-commerce systems generate huge amounts of unstructured, sequential data streams. Each sequence is a varying-length list of directional (timestamped) user-product interactions. There are hidden patterns within those sequences, users tend to interact with similar products, and based on this behavior, we can recommend the next n-products that the user may be interested in.
+The users of e-commerce systems generate huge amounts of unstructured, sequential data streams. Each sequence is a varying-length list of directional (timestamped) user-product interactions. There are hidden patterns within those sequences, users tend to interact with similar products, and based on this behavior, we can recommend the sequence of products that the user may be interested in.
 
-The `wsknn` package is a lightweight tool for modeling user-item interactions and making recommendations from sequential datasets. It is based on the k-Nearest Neighbors algorithm prepared to work with any categorical, sequential, and timestamped data, especially this generated within e-commerce systems. The package may be a stand-alone recommender, a reference against the more complex recommender systems, or a part of a bigger Machine Learning pipeline.
+The `WSKNN` (*Weighted Session-based K-Nearest Neighbors*) package is a lightweight tool for modeling user-item interactions and making recommendations from sequential datasets [@Latifi2020SessionawareRA]. It is based on the k-Nearest Neighbors algorithm (k-NN) which works with categorical, sequential, and timestamped data, especially generated in e-commerce systems. The package may be a stand-alone recommender, a reference against the more complex recommender systems, or a part of a bigger Machine Learning pipeline.
 
 # Statement of need
 
-`Wsknn` is an abbreviation from *Weighted Session-based K-NN recommender*, but algorithm itself is tuned and enhanced **Vector Multiplication Session-Based kNN (V-SKNN)** [@Ludewig2018]. The package utilizes the k-Nearest Neighbors algorithm that works with loosely structured data in the form of sequences with different lengths. This type of data is the most common representation of the timestamped events stream from a customer. The simple session example is shown below:
+`WSKNN` is an abbreviation from *Weighted Session-based k-NN recommender*, but the algorithm itself is a tuned and enhanced version of the **Vector Multiplication Session-Based kNN (V-SKNN)** algorithm [@Ludewig2018]. The package utilizes the k-NN algorithm that works with loosely structured sequential data, where sequences can have different lengths. This data type is the most common representation of the timestamped events stream from customers. A dataset example is *RecSys Challenge 2015 and the YOOCHOOSE Dataset* [@recsys2015data].
+
+The `WSKNN` recommender was designed to evaluate complex deep-learning-based architectures [@Twardowski2021], but during the research, it became clear that the k-NN model's performance is close to or exceeds the performance of the neural nets algorithms (see [experimental comparison](#experiments)). Moreover, the literature analysis about recommender systems shows that the k-NN-based solutions are performing well in different conditions [@Ludewig2018]. It makes `WSKNN` a great benchmarking tool against novel algorithms and architectures, and the first-choice tool for the fresh start and design of the recommender system.
+
+The package's algorithm can be a recommender for small and medium-sized datasets. During our own studies, the algorithm performed well for the small datasets (25k sessions; 3k items) and bigger datasets - see MovieLens 25M tutorial [@movielenstutorial25]. The model has its limitations, and the main drawback is that it is memory-hungry. As a memory-based method, it can grow up to the moment, when its usage is unfeasible. It could be an issue for production environments where the memory costs may exceed potential benefits.
+
[…]
