+++ paper/paper.md
@@ -1,20 +1,20 @@
 ---
 title: WSKNN - Weighted Session-based K-NN recommender system
 tags:
- - Python
- - machine learning
- - e-commerce
- - recommender system
- - recommender engine
+- Python
+- machine learning
+- e-commerce
+- recommender system
+- recommender engine
 authors:
- - name: Szymon Moliński
-   orcid: 0000-0003-3525-2104
-   affiliation: "1,2"
+- name: Szymon Moliński
+  orcid: 0000-0003-3525-2104
+  affiliation: "1,2"
 affiliations:
- - name: Sales Intelligence
-   index: 1
- - name: Digitree SA
-   index: 2
+- name: Sales Intelligence Sp. z o.o.
+  index: 1
+- name: Digitree SA
+  index: 2
 date: 31 March 2023
 bibliography: paper.bib
 
@@ -22,32 +22,32 @@ bibliography: paper.bib
 
 # Summary
 
-The users of e-commerce systems generate huge amounts of unstructured, sequential data streams. Each sequence is a varying-length list of directional (timestamped) user-product interactions. There are hidden patterns within those sequences, users tend to interact with similar products, and based on this behavior, we can recommend the sequence of products that the user may be interested in.
+The users of e-commerce systems generate vast amounts of unstructured, sequential data streams. Each sequence is a varying-length list of directional (timestamped) user-product interactions. There are hidden patterns within those sequences. Users tend to interact with similar products, and interactions change over time. Based on this behavior, we can recommend the sequence of products the user may be interested in.
 
-The `WSKNN` (*Weighted Session-based K-Nearest Neighbors*) package is a lightweight tool for modeling user-item interactions and making recommendations from sequential datasets [@Latifi2020SessionawareRA]. It is based on the k-Nearest Neighbors algorithm (k-NN) which works with categorical, sequential, and timestamped data, especially generated in e-commerce systems. The package may be a stand-alone recommender, a reference against the more complex recommender systems, or a part of a bigger Machine Learning pipeline.
+The `WSKNN` (*Weighted Session-based K-Nearest Neighbors*) package is a lightweight tool for modeling user-item interactions and making recommendations from sequential datasets [@Latifi2020SessionawareRA]. It is based on the k-Nearest Neighbors algorithm (k-NN), which works with categorical, sequential, and timestamped data mainly generated in e-commerce systems. The package may be a stand-alone recommender, a reference against the more complex recommender systems, or a part of a Machine Learning pipeline.
 
 # Statement of need
 
-`WSKNN` is an abbreviation from *Weighted Session-based k-NN recommender*, but the algorithm itself is a tuned and enhanced version of the **Vector Multiplication Session-Based kNN (V-SKNN)** algorithm [@Ludewig2018]. The package utilizes the k-NN algorithm that works with loosely structured sequential data, where sequences can have different lengths. This data type is the most common representation of the timestamped events stream from customers. A dataset example is *RecSys Challenge 2015 and the YOOCHOOSE Dataset* [@recsys2015data].
+`WSKNN` is an abbreviation from *Weighted Session-based k-NN recommender*. The algorithm is a tuned and enhanced version of the **Vector Multiplication Session-Based kNN (V-SKNN)** algorithm [@Ludewig2018]. The package utilizes the k-NN algorithm that works with loosely structured sequential data, where sequences can have different lengths. This data type is the most common representation of the timestamped events stream from customers. A dataset example is *RecSys Challenge 2015 and the YOOCHOOSE Dataset* [@recsys2015data].
 
[…]
