### 2023-09-27 nokaut/wsknn#51 @SimonMolinsky
Paper updates
Referring to: https://github.com/openjournals/joss-reviews/issues/5639#issuecomment-1737103021

```
line 36: it can grow up
line 39: The company owns the price comparison service Nokaut.pl
line 58: ) libraries.
line 81: Then, a dictionary with settings may be passed to the predict() function.
line 90: Extend the sentence: "The preprocessing module prepares data." with a bit more details or remove it.
line 99: The basic data type required by the algorithm is an event: (make event bold)
line 105, 107: session (make bold)
line 119: This section describes the performance of WSKNN.
line 119: from the internal
line 119: in -> at
line 120: company
line 122: A comparison
```
