+++ paper/paper.md
@@ -32,9 +32,9 @@ The `WSKNN` (*Weighted Session-based K-Nearest Neighbors*) package is a lightwei
 
 The `WSKNN` recommender was designed to evaluate complex deep-learning architectures [@Twardowski2021]. During the research, it became clear that the k-NN model's performance is close to or exceeds the performance of neural networks algorithms (see [experimental comparison](#experiments)). Moreover, the literature analysis about recommender systems shows that the k-NN-based solutions are performing well in different conditions [@Ludewig2018]. It makes `WSKNN` a great benchmarking tool against novel algorithms and architectures and the first-choice tool for the fresh start and design of the recommender system.
 
-The package's algorithm can be a recommender for small and medium-sized datasets. During the internal studies in the company, the algorithm performed well for the small datasets (25k sessions; 3k items) and bigger datasets - see MovieLens 25M tutorial [@movielenstutorial25]. The model has its limitations, and the main drawback is that it is memory-hungry. As a memory-based method, it can grow up to the moment when its usage is unfeasible. It could be an issue for production environments where the memory costs may exceed potential benefits.
+The package's algorithm can be a recommender for small and medium-sized datasets. During the internal studies in the company, the algorithm performed well for the small datasets (25k sessions; 3k items) and bigger datasets - see MovieLens 25M tutorial [@movielenstutorial25]. The model has its limitations, and the main drawback is that it is memory-hungry. As a memory-based method, it can grow to the moment when its usage is unfeasible. It could be an issue for production environments where the memory costs may exceed potential benefits.
 
-The package was created during the research project of the Sales Intelligence Sp. z o.o. company [@Twardowski2021]. The company owns the price comparison service (*Nokaut.pl*) and cooperates with multiple big stores across Poland. Thus, it has access to vast amounts of sequential data sources. Currently, the package is used for SMS and mailing recommendations for big customers.
+The package was created during the research project of the Sales Intelligence Sp. z o.o. company [@Twardowski2021]. The company owns the price comparison service *Nokaut.pl* and cooperates with multiple big stores across Poland. Thus, it has access to vast amounts of sequential data sources. Currently, the package is used for SMS and mailing recommendations for big customers.
 
 # Related work
 
@@ -44,7 +44,7 @@ The other example of a repository with scripts that is not a package is [@gru4re
 
 # Package structure
 
-The package is lightweight. It depends on the `numpy` [@harris2020array], `pandas` [@reback2020pandas], `tqdm` [@casper_da_costa_luis_2023_8233425], `more_itertools` [@moreitertools], and `pyyaml` [@pyyaml]. It works with currently supported Python versions, starting from Python 3.8. It has two main functions:
+The package is lightweight. It depends on the `numpy` [@harris2020array], `pandas` [@reback2020pandas], `tqdm` [@casper_da_costa_luis_2023_8233425], `more_itertools` [@moreitertools], and `pyyaml` [@pyyaml] libraries. It works with currently supported Python versions, starting from Python 3.8. It has two main functions:
 
 - `fit()` to build a memory representation of a model as Python dictionaries with the session-items and item-sessions maps of varying sizes.
 - `predict()` to return recommendations. It is worth noticing that **the recommendation strategy may be altered after fitting a model**; it allows testing different weighting scenarios in parallel without additional models training.
@@ -61,7 +61,7 @@ The user may pass additional parameters to the `predict()` method as a dictionar
 - is there any event (user action) that must be performed within a session to build a similarity map (for example, the *transaction* event)?
[…]
