### 2023-10-01 #5639 @AoifeHughes
Fab, thanks for the updates @SimonMolinsky can you check the current release and metadata info for the paper is correct (version numbers etc.) I'll also just generate a follow up list
