### 2023-10-01 nokaut/wsknn#54 @AoifeHughes
JOSS editor paper comments 
with respect to https://github.com/openjournals/joss-reviews/issues/5639 I have some general thoughts, but these are NOT required, please only use as guidance.

1. **Summary**
    - Change "The users of e-commerce systems generate" to "Users of e-commerce systems generate".
    - Change "Based on this behavior, we can recommend the sequence of products the user may be interested in." to "Based on this behavior, we can recommend the sequence of products that the user may be interested in."

2. **Statement of need**
    - Replace "`WSKNN` is an abbreviation from" with "`WSKNN` stands for".
    - Change "the k-NN model's performance is close to or exceeds the performance of neural networks algorithms" to "the performance of the k-NN model is comparable to, if not better than, that of neural network algorithms".
    - Change "It makes `WSKNN` a great benchmarking tool" to "This makes `WSKNN` a valuable benchmarking tool".

3. **Package structure**
    - After "It has two main functions:", consider using a colon to lead into the bullet points.
    - Use bullet points for the list of parameters under "Those parameters are:".
    - Change "The `YAML` file with documented options is provided" to "A `YAML` file documenting options is provided".

4. **Data Formats**
    - Change "The basic data type required by the algorithm is an **event**. An **event** has" to "The basic data type required by the algorithm is an **event**, which consists of:".
    - Consider using bullet points for the components of an event.

5. **Experiments**
    - Change "We see that performance on analytical metrics is close to that of RNN-based models, but the `WSKNN` model is worst regarding response times." to "While the performance of `WSKNN` on analytical metrics is comparable to RNN-based models, its response times are less optimal."

6. **Limitations**
    - Change "As with every Machine Learning system, `WSKNN` has its limitations:" to "Like all machine learning systems, `WSKNN` has limitations:".
