### 2023-10-02 nokaut/wsknn#54 @SimonMolinsky
@AoifeHughes Thank you very much! I'll go through your comments and update the paper accordingly!
