### 2023-10-02 nokaut/wsknn#54 @SimonMolinsky
@AoifeHughes I've corrected all the stylistic issues pointed out by you. Additionally, you mentioned using bullet lists, and it was a surprise for me because I had used those in places marked by you! Something may have gone wrong with the rendering. I wasn't sure, so I've changed bullets from `-` to `*`, and now everything seems to render fine in the paper.
