### 2023-10-02 #5639 @SimonMolinsky
@AoifeHughes Thank you very much! I've followed your suggestions from the issue [here](https://github.com/nokaut/wsknn/issues/54). The paper is now re-rendered and ready for the final checks :)

**Regarding metadata**

- The release has changed. Now it is `1.2.0` - I uploaded it into [PyPI](https://pypi.org/project/wsknn/), [Zenodo](https://zenodo.org/record/8398291), and set it in the GitHub repository.
