### 2023-10-06 #5639 @SimonMolinsky
> I've just seen that https://zenodo.org/record/8398291 doesn't appear to have the entire repo?

@AoifeHughes I put there only package as it is on PyPI (tar.gz and whl files). Should I upload a zipped repository (the main branch + all the files)? Should I upload it as the "Other" type of upload?
