### 2023-10-06 #5639 @AoifeHughes
You can connect a git repo to Zenodo https://coderefinery.github.io/github-without-command-line/doi/

Alternatively, yes you can upload the zip of https://github.com/nokaut/wsknn/releases if you'd rather.
