### 2023-10-06 #5639 @SimonMolinsky
Thank you for the link to the tutorial! Following it, I created a new Zenodo archive here: **https://doi.org/10.5281/zenodo.8414247**

It is linked to the GitHub release of the source code (1.2.0). Metadata is the same as before.
